+++ test/svlog/exprs/iss147_triple_binops.sv
@@ -1,0 +1,7 @@
+// RUN: moore %s -e foo
+
+module foo;
+    int a, b;
+    initial if (a === b);
+    initial if (a !== b);
+endmodule