+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Add parsing support for `assert final` assertions (#139)
 - Add parsing support for `default disable` statements (#140)
+- Add codegen support for case/wildcard equality operators (#147)
 
 ### Fixed
 - Fix sign casts (#138)